fix: some CRUD errors

+++ server/routes/statuses.js
@@ -7,50 +7,60 @@ export default (app) => {
     .get('/statuses', { name: 'page of statuses list' }, async (req, reply) => {
       if (!req.session.get('id')) {
         req.flash('error', i18next.t('flash.authError'));
-        return reply.redirect(app.reverse('root'));
+        reply.redirect(app.reverse('root'));
+        return reply;
       }
       const statuses = await app.objection.models.status.query();
-      return reply.render('statuses/index', { statuses });
+      reply.render('statuses/index', { statuses });
+      return reply;
     })
     .get('/statuses/new', { name: 'page to create status' }, async (req, reply) => {
       if (!req.session.get('id')) {
         req.flash('error', i18next.t('flash.authError'));
-        return reply.redirect(app.reverse('root'));
+        reply.redirect(app.reverse('root'));
+        return reply;
       }
       const status = new app.objection.models.status();
-      return reply.render('statuses/new', { status });
+      reply.render('statuses/new', { status });
+      return reply;
     })
     .get('/statuses/:id', { name: 'page to update status' }, async (req, reply) => {
       if (!req.session.get('id')) {
         req.flash('error', i18next.t('flash.authError'));
-        return reply.redirect(app.reverse('root'));
+        reply.redirect(app.reverse('root'));
+        return reply;
       }
       const id = +req.params?.id;
       const status = await app.objection.models.status.query().findById(id);
-      return reply.render('statuses/edit', { status, errors: {} });
+      reply.render('statuses/edit', { status, errors: {} });
+      return reply;
     })
     .post('/statuses', { name: 'create status' }, async (req, reply) => {
       if (!req.session.get('id')) {
         req.flash('error', i18next.t('flash.authError'));
-        return reply.redirect(app.reverse('root'));
+        reply.redirect(app.reverse('root'));
+        return reply;
       }
       const { data } = req.body;
       try {
         const status = await app.objection.models.status.fromJson(data);
         await app.objection.models.status.query().insert(status);
         req.flash('success', i18next.t('flash.statuses.create.success'));
-        return reply.redirect(app.reverse('page of statuses list'));
+        reply.redirect(app.reverse('page of statuses list'));
+        return reply;
       } catch (error) {
         console.log('error', error);
         req.flash('error', i18next.t('flash.statuses.create.error'));
         reply.statusCode = 422;
-        return reply.render('statuses/new', { status: data, errors: error.data });
+        reply.render('statuses/new', { status: data, errors: error.data });
+        return reply;
       }
     })
     .patch('/statuses/:id', { name: 'update status' }, async (req, reply) => {
       if (!req.session.get('id')) {
         req.flash('error', i18next.t('flash.authError'));
-        return reply.redirect(app.reverse('root'));
+        reply.redirect(app.reverse('root'));
+        return reply;
       }
       const { data } = req.body;
       const id = +req.params?.id;
@@ -58,33 +68,45 @@ export default (app) => {
         const status = await app.objection.models.status.query().findById(id);
         if (!status) {
           req.flash('error', i18next.t('flash.statuses.update.error'));
-          return reply.redirect(app.reverse('page of statuses list'));
+          reply.redirect(app.reverse('page of statuses list'));
+          return reply;
         }
         await status.$query().update(data);
         req.flash('success', i18next.t('flash.statuses.update.success'));
-        return reply.redirect(app.reverse('page of statuses list'));
+        reply.redirect(app.reverse('page of statuses list'));
+        return reply;
       } catch (error) {
         console.log('error', error);
         req.flash('error', i18next.t('flash.statuses.update.error'));
         reply.statusCode = 422;
-        return reply.render(app.reverse('page to update status', { id, status: data, errors: error.data }));
+        reply.render('statuses/edit', { status: { ...data, id }, errors: error.data });
+        return reply;
       }
     })
     .delete('/statuses/:id', { name: 'delete status' }, async (req, reply) => {
       if (!req.session.get('id')) {
         req.flash('error', i18next.t('flash.authError'));
-        return reply.redirect(app.reverse('root'));
+        reply.redirect(app.reverse('root'));
+        return reply;
       }
       const id = +req.params?.id;
       try {
+        const status = await app.objection.models.status.query().findById(id);
+        if (!status) {
+          req.flash('error', i18next.t('flash.statuses.delete.error'));
+          reply.redirect(app.reverse('page of statuses list'));
+          return reply;
+        }
         await app.objection.models.status.query().deleteById(id);
         req.flash('success', i18next.t('flash.statuses.delete.success'));
-        return reply.redirect(app.reverse('page of statuses list'));
+        reply.redirect(app.reverse('page of statuses list'));
+        return reply;
       } catch (error) {
         console.log('error', error);
         req.flash('error', i18next.t('flash.statuses.delete.error'));
         reply.statusCode = 422;
-        return reply.redirect(app.reverse('page of statuses list'));
+        reply.redirect(app.reverse('page of statuses list'));
+        return reply;
       }
     });
 };
